chore: organize navigation types (#324)
* chore: removed unnecessary comment for sharedTypes/index

* chore: copy nav screen types into types/navigation

* chore: moved custom navigation types into sharedTypes/navigation

* chore: update app stack list import

* chore: update DeviceNamingScreen import

* chore: update AppList import

* chore: update HomeTabsList import

* chore: removed InivteProps from sharedType/index

* chore: removed DeviceRole from sharedType/index

* chore: update DeviceRole import

* chore: update DeviceType import

* chore: update NativeHomeTabsNavigationProps imports

* chore: removed unused import

* chore: update NativeNavigationComponent import

* chore: update NativeRootNavigationProps import

* chore: update TabName import

* chore: updates TabBarIconProps import

* chore: device naming sceens to deviceNamingParamsList

* chore: deviceNamingParams

* chore: homeTabsList to HomeTabsParamsList

* chore: update deviceType import

* chore: update device role imports

* chore: plural of params

* chore: update AppStackList to AppStackParamsList
