+++ src/frontend/Navigation/Stack/AppScreens.tsx
@@ -1,4 +1,3 @@
-import {NavigatorScreenParams} from '@react-navigation/native';
 import * as React from 'react';
 import {RootStack} from '.';
 import {MessageDescriptor} from 'react-intl';
@@ -25,7 +24,6 @@ import {ProjectCreated} from '../../screens/Settings/CreateOrJoinProject/CreateP
 import {JoinExistingProject} from '../../screens/Settings/CreateOrJoinProject/JoinExistingProject';
 import {YourTeam} from '../../screens/Settings/ProjectSettings/YourTeam';
 import {SelectDevice} from '../../screens/Settings/ProjectSettings/YourTeam/SelectDevice';
-import {DeviceType} from '../../sharedTypes';
 import {SelectInviteeRole} from '../../screens/Settings/ProjectSettings/YourTeam/SelectInviteeRole';
 import {ReviewInvitation} from '../../screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/ReviewInvitation';
 import {InviteAccepted} from '../../screens/Settings/ProjectSettings/YourTeam/InviteAccepted';
@@ -42,7 +40,6 @@ import {
   LocationInfoScreen,
   createNavigationOptions as createLocationInfoNavOptions,
 } from '../../screens/LocationInfoScreen';
-import {InviteProps} from '../../sharedTypes';
 import {SaveTrackScreen} from '../../screens/MapScreen/track/SaveTrackScreen';
 import {InviteDeclined} from '../../screens/Settings/ProjectSettings/YourTeam/InviteDeclined';
 import {UnableToCancelInvite} from '../../screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/UnableToCancelInvite';
@@ -55,74 +52,9 @@ import {
   ManualGpsScreen,
   createNavigationOptions as createManualGpsNavigationOptions,
 } from '../../screens/ManualGpsScreen';
-import {HomeTabs, type HomeTabsList} from '../Tab';
+import {HomeTabs} from '../Tab';
 
 export const TAB_BAR_HEIGHT = 70;
-
-export type AppList = {
-  Home: NavigatorScreenParams<HomeTabsList>;
-  GpsModal: undefined;
-  Settings: undefined;
-  ProjectConfig: undefined;
-  AboutMapeo: undefined;
-  LanguageSettings: undefined;
-  CoordinateFormat: undefined;
-  Experiments: undefined;
-  PhotosModal: {
-    photoIndex: number;
-    observationId?: string;
-    editing: boolean;
-  };
-  PhotoView: undefined;
-  PresetChooser: undefined;
-  AddPhoto: undefined;
-  Observation: {observationId: string};
-  ObservationEdit: {observationId?: string} | undefined;
-  ManualGpsScreen: undefined;
-  ObservationDetails: {question: number};
-  LeaveProjectScreen: undefined;
-  AlreadyOnProj: undefined;
-  AddToProjectScreen: undefined;
-  UnableToLinkScreen: undefined;
-  ConnectingToDeviceScreen: {task: () => Promise<void>};
-  ConfirmLeavePracticeModeScreen: {projectAction: 'join' | 'create'};
-  CreateProject: undefined;
-  Security: undefined;
-  DirectionalArrow: undefined;
-  P2pUpgrade: undefined;
-  MapSettings: undefined;
-  BackgroundMaps: undefined;
-  BackgroundMapInfo: {
-    bytesStored: number;
-    id: string;
-    styleUrl: string;
-    name: string;
-  };
-  BGMapsSettings: undefined;
-  AuthScreen: undefined;
-  AppPasscode: undefined;
-  ObscurePasscode: undefined;
-  ConfirmPasscodeSheet: {passcode: string};
-  DisablePasscode: undefined;
-  SetPasscode: undefined;
-  EnterPassToTurnOff: undefined;
-  AppSettings: undefined;
-  ProjectSettings: undefined;
-  CreateOrJoinProject: undefined;
-  ProjectCreated: {name: string};
-  JoinExistingProject: undefined;
-  YourTeam: undefined;
-  SelectDevice: undefined;
-  SelectInviteeRole: {name: string; deviceType: DeviceType; deviceId: string};
-  ReviewAndInvite: InviteProps;
-  InviteAccepted: InviteProps;
-  InviteDeclined: InviteProps;
-  UnableToCancelInvite: InviteProps;
-  DeviceNameDisplay: undefined;
-  DeviceNameEdit: undefined;
-  SaveTrack: undefined;
-  Sync: undefined;
-};
 
 // **NOTE**: No hooks allowed here (this is not a component, it is a function
 // that returns a react element)
