+++ src/frontend/Navigation/Stack/index.tsx
@@ -2,13 +2,10 @@ import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack/lib/typescript/src/types';
 import {WHITE} from '../../lib/styles';
-import {AppList} from './AppScreens';
 import {CustomHeaderLeft} from '../../sharedComponents/CustomHeaderLeft';
-import {DeviceNamingSceens} from './DeviceNamingScreens';
+import {AppStackParamsList} from '../../sharedTypes/navigation';
 
-export type AppStackList = AppList & DeviceNamingSceens;
-
-export const RootStack = createNativeStackNavigator<AppStackList>();
+export const RootStack = createNativeStackNavigator<AppStackParamsList>();
 
 export const NavigatorScreenOptions: NativeStackNavigationOptions = {
   presentation: 'card',
