+++ src/frontend/Navigation/Tab/TabBar/TrackingTabBarIcon.tsx
@@ -3,7 +3,7 @@ import {StyleSheet, View} from 'react-native';
 import {TabBarIcon} from './TabBarIcon';
 import {useTracking} from '../../../hooks/tracks/useTracking';
 import {Text} from '../../../sharedComponents/Text';
-import {TabBarIconProps} from '../../../sharedTypes';
+import {TabBarIconProps} from '../../../sharedTypes/navigation';
 import {useTrackTimerContext} from '../../../contexts/TrackTimerContext';
 import {useTabNavigationStore} from '../../../hooks/useTabNavigationStore';
