+++ src/frontend/screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/ReviewInvitation.tsx
@@ -7,8 +7,8 @@ import {WHITE} from '../../../../../lib/styles';
 import {Button} from '../../../../../sharedComponents/Button';
 import {DeviceNameWithIcon} from '../../../../../sharedComponents/DeviceNameWithIcon';
 import {RoleWithIcon} from '../../../../../sharedComponents/RoleWithIcon';
-import {DeviceRole, DeviceType} from '../../../../../sharedTypes';
-import {MEMBER_ROLE_ID} from '../../../../../sharedTypes';
+import {DeviceType, MEMBER_ROLE_ID} from '../../../../../sharedTypes';
+import {DeviceRole} from '../../../../../sharedTypes';
 
 const m = defineMessages({
   title: {
